+++ packages/drawer/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [8.0.0-alpha.28](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@8.0.0-alpha.27...@react-navigation/drawer@8.0.0-alpha.28) (2026-05-08)
+
+**Note:** Version bump only for package @react-navigation/drawer
+
+# [8.0.0-alpha.27](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@8.0.0-alpha.26...@react-navigation/drawer@8.0.0-alpha.27) (2026-05-08)
+
+**Note:** Version bump only for package @react-navigation/drawer
+
 # [8.0.0-alpha.26](https://github.com/react-navigation/react-navigation/compare/@react-navigation/drawer@8.0.0-alpha.25...@react-navigation/drawer@8.0.0-alpha.26) (2026-05-05)
 
 **Note:** Version bump only for package @react-navigation/drawer